(REF) APIv4 FieldSpec - Extract DataTypeSpecTrait (dataType, serialize, fkEntity)
Also: Update `setSerialize()` to accept types a string.  This is useful if
the info comes from an annotation.
